+++ src/core/hle/kernel/timer.h
@@ -65,7 +65,7 @@ public:
      * This method should not be called from outside the timer callback handler,
      * lest multiple callback events get scheduled.
      */
-    void Signal(int cycles_late);
+    void Signal(s64 cycles_late);
 
 private:
     Timer();
